+++ sdk/test-utils/recorder-new/src/index.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export { TestProxyHttpClient } from "./core-v2-recorder";
-export { TestProxyHttpClientCoreV1 } from "./core-v1-recorder";
+export { recorderHttpPolicy, RecorderClient } from "./core-v2-recorder";
+export { TestProxyHttpClientCoreV1, RecorderClientCoreV1 } from "./core-v1-recorder";
 export { relativeRecordingsPath } from "./utils/relativePathCalculator";
 export { SanitizerOptions, RecorderStartOptions } from "./utils/utils";
 export { NoOpCredential } from "./recorderAADCredential";