+++ sdk/test-utils/recorder-new/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## 1.0.0 (Unreleased)
 
+## 2021-12-17
+
+- Refactoring the test proxy http clients for better clarity for the end users
+  - Client Renames
+    - `TestProxyHttpClient` -> `RecorderClient`
+    - `TestProxyHttpClientCoreV1` -> `RecorderClientCoreV1`
+  - Design Changes
+    - With `RecorderClientCoreV1`, instead of passing the whole client as the `httpClient` for core-v1 sdks, users are now expected to pass `recorder.httpClientCoreV1`(where `recorder` is an instantiation of `RecorderClientCoreV1`).
+    - (Internal) Abstracted the redirect/modify request logic into a new `RecorderReqeustModifier` class.
+
 ## 2021-12-15
 
 - Change the API for using variables. Variables can now be accessed using the syntax: